Camera/requirements.txt
numpy
opencv-python
pyserial
picamera2
